fix: address Copilot PR review findings for #43
- test_fov2box_time_anchor.py: add reproject_algorithm and reproject_scan
  to _make_transition_cfg params; fixes TypeError in Fov2BoxConfig.__init__
  that caused all CI tests to fail
- pyampp/io/model.py: move ensure_geometry_contract_in_metadata below the
  import block; was erroneously defined between two import statements
- real_data_checks/export_none_base_maps_to_fits.py: correct module docstring
  dtype claims (float32->float64, int16->int32) to match _COMPONENTS tuple

+++ real_data_checks/export_none_base_maps_to_fits.py
@@ -3,11 +3,11 @@
 
 Output: multi-extension FITS (MEF).
   HDU 0  – empty primary with run metadata
-  HDU 1  – BX        (base/bx,  float32,  Gauss)
-  HDU 2  – BY        (base/by,  float32,  Gauss)
-  HDU 3  – BZ        (base/bz,  float32,  Gauss)
-  HDU 4  – IC        (base/ic,  float32,  continuum intensity, normalised)
-  HDU 5  – CHROMO_MASK  (base/chromo_mask, int16)
+  HDU 1  – BX        (base/bx,  float64,  Gauss)
+  HDU 2  – BY        (base/by,  float64,  Gauss)
+  HDU 3  – BZ        (base/bz,  float64,  Gauss)
+  HDU 4  – IC        (base/ic,  float64,  continuum intensity, normalised)
+  HDU 5  – CHROMO_MASK  (base/chromo_mask, int32)
 
 WCS is taken from the stored base/index header bytes when present;
 otherwise a minimal fallback WCS is written from the geometry contract.
